fix(capture-sdk): Fix import of classes that were moved to the Bank API Library
PIA-2179
